replaced return types in IIpcCommandhandler to use ValueTask updated classess taht inherit from IIpcCommandHandler Created a temporary set on launch of uri in Daemon Worker Moved scoped db context in worker to constructor updated HttpClientBaseConfigHandler added semaphoreslim in token store changed name of Configuration of tables to DbTable in WebApi Created a bunch of tables in WebApi

+++ src/SimpleStorageSystem.Daemon/Models/Tables/FileItem.cs
@@ -14,7 +14,7 @@ public class FileItem : MetaData
     // public DateTime? LastSync { get; set; }
     // public int? PendingSyncOperation { get; set; }
 
-    public int StorageDriveId { get; set; }
+    public long StorageDriveId { get; set; }
     public StorageDrive Drive { get; set; } = null!;
     
 }
